+++ src/check.rs
@@ -352,7 +352,7 @@ impl VirtualEnv {
                         self.handle_sig(f);
                     }
                 }
-                Try => {
+                Try | Pattern => {
                     let mut f_sig = args[0].sig;
                     for handler in &args[1..] {
                         f_sig.update_outputs(|o| o.max(handler.sig.outputs()));